Reorganize the project structure: move src directory to root path
